refactor: extract common method for parsing scriptArgs
