+++ packages/builders/scripts/vats/upgrade-mintHolder.js
@@ -1,4 +1,4 @@
-import { makeHelpers } from '@agoric/deploy-script-support';
+import { makeHelpers, parseScriptArgs } from '@agoric/deploy-script-support';
 import { getManifestForUpgradingMintHolder } from '@agoric/vats/src/proposals/upgrade-mintHolder-proposal.js';
 
 const configurations = {
@@ -94,30 +94,9 @@ export const defaultProposalBuilder = async ({ publishRef, install }, opts) => {
   });
 };
 
-const Usage = `agoric run upgrade-mintHolder.js ${[...knownVariants, '<json-config>'].join(' | ')}`;
-
 /** @type {import('@agoric/deploy-script-support/src/externalTypes.js').DeployScriptFunction} */
 export default async (homeP, endowments) => {
-  const { scriptArgs } = endowments;
-  const variantOrConfig = scriptArgs?.[0];
-  console.log('upgrade-mintHolder', variantOrConfig);
-
-  const opts = {};
-
-  if (typeof variantOrConfig === 'string') {
-    if (variantOrConfig[0] === '{') {
-      try {
-        opts.config = JSON.parse(variantOrConfig);
-      } catch (err) {
-        throw Error(`Failed to parse config argument ${variantOrConfig}`);
-      }
-    } else {
-      opts.variant = variantOrConfig;
-    }
-  } else {
-    console.error(Usage);
-    throw Error(Usage);
-  }
+  const opts = parseScriptArgs(endowments, 'upgrade-mintHolder', knownVariants);
 
   const { writeCoreEval } = await makeHelpers(homeP, endowments);
   await writeCoreEval(`upgrade-mintHolder`, utils =>
